Move loading functions into loading module, make directory modules with single file into file modules
