+++ eitprocessing/sparsedata.py
